requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
Pillow
